tests/requirements.txt
selenium
nose
nose-testconfig
loguru
webdriver-manager
